% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{analyze_kingdom}
\alias{analyze_kingdom}
\title{Analyse a kingdom across document}
\usage{
analyze_kingdom(text, pages, kingdom, lang)
}
\arguments{
\item{text}{document lines (character vector)}

\item{pages}{location labels (same length as text)}

\item{kingdom}{character: "fungus", "animals", or "plants"}

\item{lang}{character: "en" or "fr"}
}
\value{
data.frame of matches (or empty)
}
\description{
Analyse a kingdom across document
}
\keyword{internal}
