% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{industrial_indicators}
\alias{industrial_indicators}
\title{Industrial indicators to exclude "plant" used in factories, sewage, etc.}
\usage{
industrial_indicators
}
\description{
Industrial indicators to exclude "plant" used in factories, sewage, etc.
}
\keyword{internal}
