% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_industrial_plant}
\alias{is_industrial_plant}
\title{Check if a context indicates industrial "plant"}
\usage{
is_industrial_plant(context)
}
\arguments{
\item{context}{character string from text}
}
\value{
logical
}
\description{
Check if a context indicates industrial "plant"
}
\keyword{internal}
