% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{make_freq_table}
\alias{make_freq_table}
\title{Make frequency table from matches}
\usage{
make_freq_table(matches, all_keywords)
}
\arguments{
\item{matches}{data.frame from search}

\item{all_keywords}{list of keyword groups (with name element)}
}
\value{
data.frame with Word and Frequency columns, plus TOTAL row
}
\description{
Make frequency table from matches
}
\keyword{internal}
